refactor(dashboard): method to config file utils
