+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/apollo/service/SentinelProjectConfigService.java
@@ -16,35 +16,25 @@
 package com.alibaba.csp.sentinel.dashboard.apollo.service;
 
 import com.alibaba.cloud.sentinel.datasource.RuleType;
-import com.alibaba.csp.sentinel.dashboard.apollo.util.DataSourceConverterUtils;
 import com.alibaba.csp.sentinel.slots.block.Rule;
-import org.apache.commons.lang.time.DateFormatUtils;
-import org.apache.tomcat.util.http.fileupload.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-import java.util.zip.ZipOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
-import static com.alibaba.csp.sentinel.dashboard.apollo.util.DataSourceConverterUtils.SERIALIZER;
+import static com.alibaba.csp.sentinel.dashboard.apollo.util.ConfigFileUtils.readProjectName2Rules;
+import static com.alibaba.csp.sentinel.dashboard.apollo.util.ConfigFileUtils.writeAsZipOutputStream;
 
 @Service
 public class SentinelProjectConfigService {
 
     private static final Logger logger = LoggerFactory.getLogger(SentinelProjectConfigService.class);
-
-    /**
-     * @see <a href="https://stackoverflow.com/questions/13846000/file-separators-of-path-name-of-zipentry">File separators of Path name of ZipEntry?</a>
-     */
-    private static final String FORWARD_SLASH = "/";
-
-    private static final String FILENAME_SUFFIX = ".json";
 
     private final SentinelApolloService sentinelApolloService;
 
@@ -52,77 +42,6 @@ public class SentinelProjectConfigService {
         this.sentinelApolloService = sentinelApolloService;
     }
 
-    static ZipEntry resolveZipEntry(String projectName, RuleType ruleType) {
-        return new ZipEntry(String.join(FORWARD_SLASH, projectName, ruleType.name() + FILENAME_SUFFIX));
-    }
-
-    /**
-     * @throws IllegalArgumentException if zipEntry is a directory
-     * @see #resolveZipEntry(String, RuleType)
-     */
-    static String getProjectName(ZipEntry zipEntry) {
-        Assert.isTrue(! zipEntry.isDirectory(), "zip entry" + zipEntry + " should not be a directory");
-        File file = new File(zipEntry.getName());
-        return file.getParent();
-    }
-
-    /**
-     * @see #resolveZipEntry(String, RuleType)
-     */
-    static RuleType getRuleType(ZipEntry zipEntry) {
-        Assert.isTrue(! zipEntry.isDirectory(), "zip entry" + zipEntry + " should not be a directory");
-        String filename = new File(zipEntry.getName()).getName();
-        Assert.isTrue(filename.endsWith(FILENAME_SUFFIX), "filename '" + filename + "' should ends with " + FILENAME_SUFFIX);
-
-        // delete suffix
-        String ruleTypeName = filename.substring(0, filename.length() - FILENAME_SUFFIX.length());
-        return RuleType.valueOf(ruleTypeName);
-    }
-
-    private static void write2ZipOutputStream(ZipOutputStream zipOutputStream, String projectName, RuleType ruleType, List<? extends Rule> rules) throws IOException {
-        ZipEntry zipEntry = resolveZipEntry(projectName, ruleType);
-
-        try {
-            zipOutputStream.putNextEntry(zipEntry);
-            String jsonContent = SERIALIZER.convert(rules);
-            zipOutputStream.write(jsonContent.getBytes(StandardCharsets.UTF_8));
-            zipOutputStream.closeEntry();
-        } catch (IOException e) {
-            String message = "config export failed. project name = " + projectName;
-            logger.error(message, projectName);
-            throw new IOException(message, e);
-        }
-    }
-
-    private static void write2ZipOutputStream(ZipOutputStream zipOutputStream, String projectName, Map<RuleType, List<? extends Rule>> ruleTypeListMap)
-            throws IOException {
-        for (Map.Entry<RuleType, List<? extends Rule>> entry : ruleTypeListMap.entrySet()) {
-            RuleType ruleType = entry.getKey();
-            List<? extends Rule> rules = entry.getValue();
-            if (rules.size() > 0) {
-                write2ZipOutputStream(zipOutputStream, projectName, ruleType, rules);
-            }
-        }
-    }
-
-    private static void writeAsZipOutputStream(OutputStream outputStream, Map<String, Map<RuleType, List<? extends Rule>>> projectName2rules)
-            throws IOException {
-        try (ZipOutputStream zipOutputStream = new ZipOutputStream(outputStream)) {
-            for (Map.Entry<String, Map<RuleType, List<? extends Rule>>> entry : projectName2rules.entrySet()) {
-                String projectName = entry.getKey();
-                Map<RuleType, List<? extends Rule>> ruleTypeListMap = entry.getValue();
-                if (ruleTypeListMap.size() > 0) {
-                    write2ZipOutputStream(zipOutputStream, projectName, ruleTypeListMap);
-                }
-            }
-        }
-    }
-
-    public static String generateZipFilename() {
-        // must contain the information of time
-        final String zipFilename = "sentinel_config_export_" + DateFormatUtils.format(new Date(), "yyyy_MMdd_HH_mm_ss") + ".zip";
-        return zipFilename;
-    }
 
     public void exportAllToZip(OutputStream outputStream) throws IOException {
         Map<String, Map<RuleType, List<? extends Rule>>> projectName2rules = this.sentinelApolloService.getRules();
@@ -143,44 +62,8 @@ public class SentinelProjectConfigService {
         writeAsZipOutputStream(outputStream, projectName2rules);
     }
 
-    private static List<? extends Rule> readRules(InputStream inputStream, RuleType ruleType) throws IOException {
-        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-        IOUtils.copy(inputStream, byteArrayOutputStream);
-        byte[] bytes = byteArrayOutputStream.toByteArray();
-        return DataSourceConverterUtils.deserialize(bytes, ruleType);
-    }
-
-    private static Map<String, Map<RuleType, List<? extends Rule>>> readFromZipInputStream(ZipInputStream zipInputStream) throws IOException {
-        Map<String, Map<RuleType, List<? extends Rule>>> projectName2rules = new HashMap<>();
-        for (
-                ZipEntry zipEntry = zipInputStream.getNextEntry();
-                null != zipEntry;
-                zipEntry = zipInputStream.getNextEntry()
-        ) {
-            if (!zipEntry.isDirectory()) {
-                String projectName = getProjectName(zipEntry);
-                if (!projectName2rules.containsKey(projectName)) {
-                    projectName2rules.put(projectName, new HashMap<>());
-                }
-
-                RuleType ruleType = getRuleType(zipEntry);
-                // read content
-                List<? extends Rule> rules = readRules(zipInputStream, ruleType);
-                // add rules
-                projectName2rules.get(projectName).put(ruleType, rules);
-            }
-
-            zipInputStream.closeEntry();
-        }
-
-        return projectName2rules;
-    }
-
     public Map<String, Map<RuleType, List<? extends Rule>>> importAllFrom(InputStream inputStream) throws IOException {
-        final Map<String, Map<RuleType, List<? extends Rule>>> projectName2rules;
-        try (ZipInputStream zipInputStream = new ZipInputStream(inputStream)) {
-            projectName2rules = readFromZipInputStream(zipInputStream);
-        }
+        final Map<String, Map<RuleType, List<? extends Rule>>> projectName2rules = readProjectName2Rules(inputStream);
 
         // registry projects
         logger.info("import {} projects config. project names = {}", projectName2rules.size(), projectName2rules.keySet());
