+++ sentinel-dashboard/src/test/java/com/alibaba/csp/sentinel/dashboard/apollo/util/ConfigFileUtilsTest.java
@@ -1,32 +1,32 @@
-package com.alibaba.csp.sentinel.dashboard.apollo.service;
+package com.alibaba.csp.sentinel.dashboard.apollo.util;
 
 import com.alibaba.cloud.sentinel.datasource.RuleType;
 import junit.framework.TestCase;
 
 import java.util.zip.ZipEntry;
 
-public class SentinelProjectConfigServiceTest extends TestCase {
+public class ConfigFileUtilsTest extends TestCase {
 
     public void testResolveZipEntry() {
-        ZipEntry zipEntry = SentinelProjectConfigService.resolveZipEntry("app", RuleType.SYSTEM);
+        ZipEntry zipEntry = ConfigFileUtils.resolveZipEntry("app", RuleType.SYSTEM);
         assertTrue(zipEntry.getName().startsWith("app"));
         assertTrue(zipEntry.getName().contains(RuleType.SYSTEM.name()));
         assertFalse(zipEntry.getName().contains(RuleType.SYSTEM.getName()));
 
-        ZipEntry zipEntry1 = SentinelProjectConfigService.resolveZipEntry("app1", RuleType.PARAM_FLOW);
+        ZipEntry zipEntry1 = ConfigFileUtils.resolveZipEntry("app1", RuleType.PARAM_FLOW);
         assertTrue(zipEntry1.getName().contains(RuleType.PARAM_FLOW.name()));
         assertFalse(zipEntry1.getName().contains(RuleType.PARAM_FLOW.getName()));
     }
 
     public void testGetProjectName() {
-        ZipEntry zipEntry = SentinelProjectConfigService.resolveZipEntry("app", RuleType.FLOW);
-        String projectName = SentinelProjectConfigService.getProjectName(zipEntry);
+        ZipEntry zipEntry = ConfigFileUtils.resolveZipEntry("app", RuleType.FLOW);
+        String projectName = ConfigFileUtils.getProjectName(zipEntry);
         assertEquals("app", projectName);
     }
 
     public void testGetRuleType() {
-        ZipEntry zipEntry = SentinelProjectConfigService.resolveZipEntry("app", RuleType.AUTHORITY);
-        RuleType ruleType = SentinelProjectConfigService.getRuleType(zipEntry);
+        ZipEntry zipEntry = ConfigFileUtils.resolveZipEntry("app", RuleType.AUTHORITY);
+        RuleType ruleType = ConfigFileUtils.getRuleType(zipEntry);
         assertEquals(ruleType, RuleType.AUTHORITY);
     }
